requirements.txt
colour-science==0.4.6
imageio==2.37.0
numpy==2.3.0
pillow==11.2.1
scipy==1.15.3
typing_extensions==4.14.0
